python/requirements.txt
langchain
chromadb
pypdf
pytest
langchain-community
langchain-ollama
langchain-chroma
pytesseract
pillow
python-docx
python-pptx
openpyxl
pandas
transformers
fpdf
PyMuPDF
aiofiles
chardet